requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
qrcode
python-dotenv
pillow 
sendgrid
